README.md수정, CarRacer -> CarNavigator 네이밍변경, MoveCars기능, MoveCarIfPossible 기능 추가
CarNavigator가 책임을 더 잘표현한다고 생각하여 네이밍 변경하였습니다.
MoveCars기능은 Car의 리스트를 토대로 루프를 돌면서 차가 움직이는것이 가능하면 차를움직이는 메소드를 호출합니다..
MoveCarIfPossible은 차가 움직일수있는지를 판단하는 메소드를 호출하고 가능하마녀 차하나를 움직이는 메소드를 호출하는 메소드입니다.
ReadMe 에 MoveCars, MoveCarIfPossible 기능을 추가하고 CarNavigator네이밍을 변경하였습니다.
